web/requirements.txt
Django==1.5.4
MySQL-python==1.2.2
PAM==0.4.2
Twisted-Core==10.1.0
Twisted-Names==10.1.0
Twisted-Web==10.1.0
beautifulsoup4==4.3.2
distribute==0.6.14
foolscap==0.5.0
ipython==0.10
matplotlib==0.99.3
numpy==1.4.1
pexpect==2.3
pyOpenSSL==0.10
pyparsing==1.5.2
pyserial==2.3
python-apt==0.7.100.1-squeeze1
python-dateutil==1.4.1
pytz==2010b
reportbug==4.12.6
scipy==0.7.2
wsgiref==0.1.2
wxPython==2.8.10.1
wxPython-common==2.8.10.1
zope.interface==3.5.3
django-debug-toolbar==0.10.2
git+git://github.com/pythonforfacebook/facebook-sdk.git#egg=facebook-python-sdk
